// onecx-chat-ui/src/app/shared/generated/model/chat.ts
/**
 * onecx-chat bff
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Participant } from './participant';
import { ChatType } from './chatType';


export interface Chat { 
    version?: number;
    creationDate?: string;
    creationUser?: string;
    modificationDate?: string;
    modificationUser?: string;
    id?: string;
    type: ChatType;
    topic?: string;
    summary?: string;
    appId?: string;
    participants?: Array<Participant>;
}
export namespace Chat {
}