// onecx-chat-ui/src/app/shared/generated/model/participant.ts
/**
 * onecx-chat bff
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ParticipantType } from './participantType';


export interface Participant { 
    version?: number;
    creationDate?: string;
    creationUser?: string;
    modificationDate?: string;
    modificationUser?: string;
    id?: string;
    type: ParticipantType;
    userId: string;
    userName?: string;
    email?: string;
}
export namespace Participant {
}